+++ QtUI/views/dateSelectionFrame.py
@@ -4,7 +4,7 @@ from PyQt6.QtCore import pyqtSignal
 from Core.dataAccess.dataManager import getData_API
 from QtUI.views.rawUI.ui_rawDateSelectionFrame import Ui_dateSelection
 
-from PyQt6.QtCore import QEvent,Qt
+from PyQt6.QtCore import QDate
 
 
 
@@ -19,9 +19,11 @@ class DateSelectionFrame(QWidget):
         #  ------ 初始化 ------
         #  --- 输入框 ---
         self.dateSelectionFrame.dateTimeEdit.setDisplayFormat("yyyy-MM-dd")
+        self.dateSelectionFrame.dateTimeEdit.setDate(QDate.currentDate())
         
         #  ------ 绑定信号 ------
         self.dateSelectionFrame.dateTimeEdit.dateChanged.connect(self._on_dateTimeEdit_enter)
+        self.dateSelectionFrame.newRecordButton.clicked.connect(self._on_dateTimeConfirm)
     
     
     # def _on_enter_entered(self,obj,event):
@@ -43,8 +45,6 @@ class DateSelectionFrame(QWidget):
         
     def _on_dateTimeEdit_enter(self,q_date):
         date = q_date.toString("yyyy-MM-dd")
-        month = q_date.toString("MM")
-        day = q_date.toString("dd")
         data = getData_API("Data/dateData.json")
         
         #  --- 判断用户是否输入完成 ---
@@ -54,4 +54,14 @@ class DateSelectionFrame(QWidget):
         #  --- 传递信号 ---
         if date in data:
             self.dateSelected.emit(date)
+    
+    def _on_dateTimeConfirm(self):
+        date = self.dateSelectionFrame.dateTimeEdit.text()
+        q_date = QDate.fromString(date,"yyyy-MM-dd")
         
+        #  --- 判断是否valid ---
+        if not q_date.isValid():
+            return
+        
+        #  --- 传递信号 ---
+        self.dateSelected.emit(date)